% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{st_set_crs}
\alias{st_set_crs}
\alias{st_setsrid}
\title{Spatial Reference System}
\usage{
st_set_crs(x, value)

st_setsrid(x, value)
}
\arguments{
\item{x}{numeric, character, or object of class \link[sf]{sf} or \link[sf]{sfc}}

\item{value}{one of (i) character: a valid proj4string (ii) integer, a valid EPSG value (numeric), or (iii) a list containing named elements \code{proj4string} (character) and/or \code{epsg} (integer) with (i) and (ii).}
}
\description{
Spatial Reference System
}
\details{
\code{st_setsrid} is postgis compatible
}
