% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotidy-package.R
\docType{package}
\name{geotidy-package}
\alias{geotidy}
\alias{geotidy-package}
\title{Tidy spatial data
\code{geotidy} uses tibbles to manipulate simple featues. It uses the \link{sf} package under the hood. There are a few minor differences, but if you're already using a tidy workflow with \code{sf}, most of the code is compatible already. \code{geotidy} uses a subset of \code{sf}.}
\description{
Manipulate spatial data tidily. `Geotidy` provides a selection of 
    spatial functions from the `sf` package that are adapted to a tidy workflow. 
    It relies on tibbles and `dplyr` verbs, and forces you to be explicit about 
    the spatial operations you desire.
}
\author{
\strong{Maintainer}: Etienne Racine \email{etiennebr@gmail.com} (\href{https://orcid.org/0000-0003-1109-894X}{ORCID})

}
\keyword{internal}
