% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sfc.R
\name{map_sfc}
\alias{map_sfc}
\alias{map2_sfc}
\title{Apply a function to each element of a vector and bind to an sfc object}
\usage{
map_sfc(.x, .f, ..., .crs = NA_crs_, .precision = 0, .check_ring_dir = FALSE)

map2_sfc(
  .x,
  .y,
  .f,
  ...,
  .crs = NA_crs_,
  .precision = 0,
  .check_ring_dir = FALSE
)
}
\arguments{
\item{.x}{Vectors of the same length. A vector of length 1 will
be recycled.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.crs}{Coordinate Reference System. See \code{\link[sf:st_crs]{sf::st_crs()}} for details.}

\item{.precision}{Numeric. Precision applied to every coordinate, expressed
as a scale factor. For instance 1000 to round to three decimals.}

\item{.check_ring_dir}{Logical. Check that the polygon ring directions are
counter clockwise and holes are clockwise, order is correct it if not.
Default to `FALSE``.}

\item{.y}{Vectors of the same length. A vector of length 1 will
be recycled.}
}
\description{
Apply a function to each element of a vector and bind to an sfc object
}
\seealso{
\code{\link[purrr:map]{purrr::map()}} \link[sf:st_sfc]{sf::st_sfc}
}
