% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_dumppoints}
\alias{st_dumppoints}
\title{Dump vertex to a nested tibble of points}
\usage{
st_dumppoints(x)
}
\arguments{
\item{x}{Geometry \code{sfc} column}
}
\value{
A list of tibbles.
}
\description{
Creates a geometry column containing a tibble where each vertex is a row.
\code{st_dumppoints} is useful for expanding geometries. It is the reverse of a GROUP BY
in that it creates new rows. For example it can be used to expand
MULTIPOLYGONS into POLYGONS.
}
\seealso{
\code{\link[sf:st_cast]{sf::st_cast()}}, \code{\link[=st_coordinates]{st_coordinates()}}
}
