% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{st_envelope}
\alias{st_envelope}
\title{Create a geometry representing the bounding box of the geometry}
\usage{
st_envelope(geom)
}
\arguments{
\item{geom}{an \code{sfc} geometry.}
}
\description{
Create a geometry representing the bounding box of the geometry
}
\details{
This function does not aggregate by default (unlike \code{sf::st_bbox}), you will need \code{sf::st_combine()} to aggregate first.
}
\examples{
library(tibble)
library(dplyr)

x <- tibble(
       line = c(
         st_makeline(1:10, 11:20),
         st_makeline(1:10+2, 11:20 -2)
       ) \%>\%
       sf::st_set_crs(4326)
)

mutate(x,
  envelope = st_envelope(line)
)
# aggregate
summarise(x,
  envelope = line \%>\% sf::st_combine() \%>\% st_envelope()
)
}
\seealso{
\code{sf::st_bbox()}
}
