% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomfromtext.R
\name{st_geomfromtext}
\alias{st_geomfromtext}
\title{Geometry from text}
\usage{
st_geomfromtext(x, crs = sf::NA_crs_)
}
\arguments{
\item{x}{WKT description}

\item{crs}{coordinate reference system}
}
\value{
A geometry (class \code{sfc})
}
\description{
Construct a geometry from OGC Well-Known text (WKT) description.
}
\details{
It can be used as an equivalent to type casting in postgis (\code{'POINT(1 1)'::geometry})
}
\examples{
st_geomfromtext('POLYGON((38 16,38 50,65 50,66 16,38 16))')
st_geomfromtext('POINT(-71.064544 42.28787)', crs = 4326)
# EWKT works as well
st_geomfromtext('SRID=4326;POINT(-71.064544 42.28787)')
}
