% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{st_union}
\alias{st_union}
\title{Union geometries}
\usage{
st_union(x, y)
}
\arguments{
\item{x}{A geometry or a set of geometries of class \code{sfc} to be unioned.}

\item{y}{An optional geometry or a set of geometries of class \code{sfc}.}
}
\description{
Union geometries to create MULTI* geometries in the case of different
geometries. If geometries are identical, the union returns a single (MULTI)
geometry.
If you need to preserve all geometries, use \code{\link[sf:st_combine]{sf::st_combine()}}.
}
\details{
This function mimics postgis behavior, which is differently than \code{sf} because
it strictly returns 1 or \code{length(x)} geometries.

When provided two elements, each element of \code{x} is unioned with
corresponding \code{y} element. When provided only \code{x}, then all the rows
are unioned together to provide a single geometry.
}
\examples{
library(tibble)
library(dplyr)

geom <- tibble(
  x = 1:3,
  y = 4:6,
  origin = st_point(x, y),
  destin = st_point(x + 1, y + 1)
)
geom <- mutate(geom, union = st_union(origin, destin))
geom
summarise(geom, union = st_union(origin, destin))
}
\seealso{
\code{\link[sf:st_union]{sf::st_union()}}, \code{\link[sf:st_combine]{sf::st_combine()}}
}
