% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_coordinates}
\alias{st_coordinates}
\title{List of vertex coordinates}
\usage{
st_coordinates(x)
}
\arguments{
\item{x}{Geometry \code{sfc} column}
}
\value{
A list of tibbles.
}
\description{
List of vertex coordinates
}
\details{
\code{st_coordinates} returns a tibble containing coordinates, also some
grouping features. \code{POINT}: \code{.path = 1}; \code{MULTIPOINT} and \code{LINESTRING}: \code{.path} orders the
points; \code{LINESTRING}, \code{MULTILINESTRING} and \code{POLYGON}: \code{.l_} provides feature differentiation.
}
\seealso{
\code{\link[sf:st_cast]{sf::st_cast()}}, \code{\link[=st_dumppoints]{st_dumppoints()}}, \code{\link[sf:st_coordinates]{sf::st_coordinates()}}
}
