% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_makeline}
\alias{st_makeline}
\title{Creates a Linestring from point, multipoint, or line geometries.}
\usage{
st_makeline(.geom, .to, ...)
}
\arguments{
\item{.geom}{A geometry or a set of geometries of class \code{sfc} to be converted
to line. The geometries are often aggregated using a \code{\link[dplyr:group_by]{dplyr::group_by()}}
followed by a\code{\link[dplyr:summarise]{dplyr::summarise()}}.}

\item{.to}{A point geometry to create pairwise lines.}

\item{...}{Unused.}
}
\value{
A line (\code{LINESTRING}) of class \code{sfc}.
}
\description{
Lines can be created from point aggregates (using \code{summarise}) or from -- to
to create a 2-vertex line.
}
\examples{
library(dplyr)
library(tibble)

x <- tibble(g = c("a", "a"), point = c(st_point(12, 21), st_point(21, 12)))
x \%>\%
 summarise(line = st_makeline(point))
}
\seealso{
Other make: 
\code{\link{st_makepoint}()},
\code{\link{st_multi}()},
\code{\link{st_numpoints}()}
}
\concept{make}
