% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_multi}
\alias{st_multi}
\title{Aggregate geometries}
\usage{
st_multi(.geom, ...)
}
\arguments{
\item{.geom}{A geometry or a set of geometries of class \code{sfc} to be grouped. The geometries
are often aggregated using a \code{\link[dplyr:group_by]{dplyr::group_by()}} followed by a
\code{\link[dplyr:summarise]{dplyr::summarise()}}.}

\item{...}{Unused.}
}
\description{
Returns the geometry as a \code{MULTI*} geometry. If the geometry is already a \code{MULTI*}, it is returned unchanged.
}
\seealso{
Other make: \code{\link{st_makepoint}}
}
\concept{make}
