% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_multi}
\alias{st_multi}
\title{Aggregate geometries}
\usage{
st_multi(.geom, ...)
}
\arguments{
\item{.geom}{A geometry or a set of geometries of class \code{sfc} to be grouped. The geometries
are often aggregated using a \code{\link[dplyr:group_by]{dplyr::group_by()}} followed by a
\code{\link[dplyr:summarise]{dplyr::summarise()}}.}

\item{...}{Unused.}
}
\value{
a single \verb{MULTI*} geometry.
}
\description{
Returns the geometry as a \verb{MULTI*} geometry. If the geometry is already a \verb{MULTI*}, it is returned unchanged.
}
\examples{
library(tibble)
library(dplyr)

tibble(g = c("a", "a"), point = c(st_point(12, 21), st_point(21, 12))) \%>\%
  group_by(g) \%>\%
  summarise(multi = st_multi(point))
}
\seealso{
Other make: 
\code{\link{st_makeline}()},
\code{\link{st_makepoint}()},
\code{\link{st_numpoints}()}
}
\concept{make}
