% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{st_makepoint}
\alias{st_makepoint}
\alias{st_point}
\title{Create points}
\usage{
st_makepoint(x, y)

st_point(x, y)
}
\arguments{
\item{x}{Numeric vector of x coordinates (e.g. longitude) or a list of pairs
of coordinates. In the latter case \code{y} can be missing.}

\item{y}{Numeric vector of y coordinates (e.g. latitude).}
}
\value{
List column of points with the same length than \code{x} and \code{y}.
}
\description{
Create points from coordinates. \code{st_point} is simply a wrapper around \code{st_makepoint}.
}
\details{
\code{x} amd \code{y} must be of the same length. If one of the coordiantes are
missing, it creates and empty point geometry.
}
\examples{
library(tibble)
library(dplyr)

df <- tibble(longitude = -71.1043, latitude = 42.3150)
df \%>\%
  mutate(geometry = st_point(longitude, latitude))
}
\seealso{
\link[sf:st_point]{sf::st_point}

Other make: 
\code{\link{st_makeline}()},
\code{\link{st_multi}()},
\code{\link{st_numpoints}()}
}
\concept{make}
