% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{st_union}
\alias{st_union}
\title{Union geometries}
\usage{
st_union(x, y, by_feature = TRUE)
}
\arguments{
\item{x}{A geometry or a set of geometries of class \code{sfc} to be unioned.}

\item{y}{(optional). A geometry or a set of geometries of class \code{sfc}.}

\item{by_feature}{default is \code{TRUE}. Union each element of \code{x} with
corresponding \code{y} element. If \code{FALSE}, all the features are first unioned
row-wise, then all unioned together.}
}
\description{
Union geometries to create MULTI* geometries in the case of different
geometries. If geometries are identical, the union returns a single geometry.
If you need to preserve all geometries, use \code{\link[sf:st_combine]{sf::st_combine()}}. This function
behaves differently than \code{sf} because it strictly returns 1 or \code{length(x)}
geometries.
}
\examples{
library(tibble)
library(dplyr)

geom <- tibble(
  x = 1:3,
  y = 4:6,
  origin = st_point(x, y),
  destin = st_point(x + 1, y + 1)
)
geom <- mutate(geom, union = st_union(origin, destin))
geom
summarise(geom, union = st_union(origin, destin, by_feature = FALSE))
}
\seealso{
\code{\link[sf:st_union]{sf::st_union()}}, \code{\link[sf:st_combine]{sf::st_combine()}}
}
